dataset_stream/requirements.txt
fastapi[standard]~=0.115.12
sqlalchemy
psycopg2-binary
python-dotenv
lark
numpy
geoalchemy2
